feat(deployer): add deployment wait timeout configuration and enhance deployment handling
- Introduced DeploymentWaitTimeout field in ServiceConfig for customizable deployment wait duration.
- Updated NewServiceConfigFromConfig to initialize the new timeout setting.
- Enhanced deployment service to utilize the configured wait timeout when waiting for deployments to be ready.
- Added tests to validate the new wait functionality and ensure proper timeout handling.
